requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
pika==1.3.2
sqlparse==0.5.1
drf-yasg>=1.21.3  
prometheus-client==0.19.0
django-prometheus==2.3.1

